Update esp-presence-2450.yaml
You can define an angle if you have not placed the sensor flat against a wall
The maximum detection length has increased from 6 to 8 meters and can now start at - 0.5 meters
The positioning of the zones no longer works with the same system. Before there was a minimum X and Y zone and a maximum X and Y zone. Now there is just an X and Y position with a length and width
Each sensor entity now has the name of the room in which it is integrated. You must choose the name of this room manually. (entity_name: "Room")
To support all these changes, the code had to be written in depth with additions, deletions and modifications. It is now necessary to have an additional file next to our main YAML file. It is a code library.